requirements.txt
requests
tqdm
PyYAML
python-dotenv
unidecode
aria2p
colorama
requests-cache
requests-toolbelt
beautifulsoup4
click
click-completion
click-default-group
click-help-colors
click-log
click-spinner
click-plugins
click-repl
click-shell
click-terminator
click-plugins
click-argcomplete